change: Jumbotron to Foreign Key
Jumbotron can now have many image objects. It was initially made one to one so that jumbotron is expected to be assigned to one unique image object to avoid having a jumbotron with the same image file. But user can just upload an image object with the same image file, so jumbotron with similar files would still exist.