feat: changed model requirements; removed null on required fields
Fields such as title or description should not be made null for content such as events or projects, because these are essential parts of that content. The model serializer also does validation according to what you had explicitly required in the model, so this way, when an event model without a title is sent (in the event that form validation in the frontend failed/was unable to implement the requirement), then the backend server would go and reject that request to add data to preserve the database. An error code would be sent instead to inform the user that the data sent is invalid.