feat: added date validators for events and projects
The added validators simply check if the start date is not greater than the end date. This is used under the validation for serializers. When a post request is made, the serializer takes care of the checking of the fields, and that would include data validation. This ensures less bugs in the case that a user would input a start date greater than that of its end date.