feat: read and write serializers
read serializers focuses on displaying the data, showing all nested relationships depending on what fields are being expanded. Write serializers on the other hand encompasses of the fields needed/allowed when posting data to that endpoint. Validation logic and overrides on the create and update methods are put in this serializer.